@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
-		<Version>4.0.0</Version>
+		<TargetFramework>net8.0</TargetFramework>
+		<Version>4.1.0</Version>
 		<PackageProjectUrl>https://github.com/StrutTower/Repository</PackageProjectUrl>
 		<RepositoryUrl>https://github.com/StrutTower/Repository</RepositoryUrl>
 		<PackageIcon>towersoft-logo_128x.png</PackageIcon>
@@ -26,7 +26,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Npgsql" Version="8.0.3" />
+		<PackageReference Include="Npgsql" Version="9.0.4" />
 	</ItemGroup>
 
 	<ItemGroup>
